Changes per blue's spec
All except: `Also: how difficult would it be to make the mouse cursor go 50 pixels diagonal towards the top-right corner and then 150 pixels down from there (as in: figure 1)? If this is not too much trouble one could use the same drawings for exiting the Hz programming mode: figure 8 for 8k, figure 4 for 4k, a z for 2k and figure 1 for 1k.  `